Rename profile redesign classes
