+++ app/javascript/mastodon/features/account_timeline/components/fields.tsx
@@ -23,7 +23,7 @@ import { cleanExtraEmojis } from '../../emoji/normalize';
 import type { AccountField } from '../common';
 import { useFieldHtml } from '../hooks/useFieldHtml';
 
-import classes from './redesign.module.scss';
+import classes from './styles.module.scss';
 
 const verifyMessage = defineMessage({
   id: 'account.link_verified_on',
